Updates for Tweets JSON
Added some JSOn arrays for the Tweets, and made it so that they are
getting added into the sendmessage value so that it will be getting
passed into the chromecast app so that I can retrive it. this is mainly
done from within the TickerACtivity.java. Most other java files are now
pointless and will eventually be cleared. Also fixed a big that would
cause no page to be shown on a new install as the app id did not exist
in preferences.. This is no longer needed to be in preferences so we are
referencing the string resource appid again.